// src/core/interfaces/control.interface.ts
export interface IControl {
  url: string;
  id: number;
  __unicode__: string;
  format: number;
  date: string;
  start: string;
  subgroup: number | null;
  theme: number;
  housing: number;
  room: number;
  group_semester: number;
  teachers: number[];
  conduct_type: string;
  link: string;
}